Fixes for:
- Issue #55: warning in cub/device/dispatch/dispatch_reduce_by_key.cuh 

- Issue #59: cub::DeviceScan::ExclusiveSum can't prefix sum of float
into double

Former-commit-id: 30a63fa8a01f1ffc036a14e8c06c9b7abf53aad0

+++ cub/device/dispatch/dispatch_histogram.cuh
@@ -572,10 +572,10 @@ struct DipatchHistogram
 
             // Get grid dimensions, trying to keep total blocks ~histogram_sweep_occupancy
             int pixels_per_tile     = histogram_sweep_config.block_threads * histogram_sweep_config.pixels_per_thread;
-            int tiles_per_row       = (num_row_pixels + pixels_per_tile - 1) / pixels_per_tile;
+            int tiles_per_row       = int(num_row_pixels + pixels_per_tile - 1) / pixels_per_tile;
             int blocks_per_row      = CUB_MIN(histogram_sweep_occupancy, tiles_per_row);
             int blocks_per_col      = (blocks_per_row > 0) ?
-                                        CUB_MIN(histogram_sweep_occupancy / blocks_per_row, num_rows) :
+                                        int(CUB_MIN(histogram_sweep_occupancy / blocks_per_row, num_rows)) :
                                         0;
             int num_threadblocks    = blocks_per_row * blocks_per_col;
